+++ views/office.ejs
@@ -61,6 +61,6 @@
   <script src='https://meet.jit.si/external_api.js'></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.0.4/socket.io.js"></script>
   <script src="https://apis.google.com/js/platform.js" async defer></script>
-  <script src="./dist/office-bundle.js"></script>
+  <script src="dist/office-bundle.js"></script>
 </body>
 </html>
